backend/requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
Werkzeug==2.0.1
osmnx==0.16.1 
geopandas==0.9.0
shapely==1.8.5
kaleido
networkx
plotly
pandas